Separate dedicated bluez directory for linux BLE interfacing (#4846)
* Add an adapter iterator class to bluez

* code review updates

* Fix typo in method

* Fix compilation and off-by-one error

* Ensure we unref the adapter during listing

* Remove unused variable

* Fix typos in comments - I clobbered the fixes that Justin pushed

* Fix typos in comments - I clobbered the fixes that Justin pushed

* Add support for native adapter listing

* Restyle fixes

* Update the init logic

* Do not auto-import GetAdapters. ble is a stand alone package for now

* Update typing

* Move iterator values to std::string and fix typo in linux impl

* Move Bluez files into a separate directory to be able to better split things

* Split out AdapterIterator from bluez

* Restyle fixes

* Remove adapteriterator from helper.h and only keep in adapteriterator.h

* Switch reinterpret cast to static cast
