+++ src/platform/BUILD.gn
@@ -441,8 +441,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Linux/BLEManagerImpl.cpp",
         "Linux/BLEManagerImpl.h",
         "Linux/BlePlatformConfig.h",
-        "Linux/CHIPBluezHelper.cpp",
-        "Linux/CHIPBluezHelper.h",
         "Linux/CHIPDevicePlatformConfig.h",
         "Linux/CHIPDevicePlatformEvent.h",
         "Linux/CHIPLinuxStorage.cpp",
@@ -466,6 +464,11 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Linux/PosixConfig.h",
         "Linux/SystemPlatformConfig.h",
         "Linux/SystemTimeSupport.cpp",
+        "Linux/bluez/AdapterIterator.cpp",
+        "Linux/bluez/AdapterIterator.h",
+        "Linux/bluez/Helper.cpp",
+        "Linux/bluez/Helper.h",
+        "Linux/bluez/Types.h",
       ]
 
       if (chip_enable_mdns) {
