+++ src/platform/Linux/bluez/Types.h
@@ -1,6 +1,6 @@
 /*
  *
- *    Copyright (c) 2020 Project CHIP Authors
+ *    Copyright (c) 2021 Project CHIP Authors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -43,12 +43,9 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- *    @file
- *          Provides Bluez dbus implementation for BLE
- */
-
 #pragma once
+
+#include <platform/CHIPDeviceLayer.h>
 
 #if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
 
@@ -182,88 +179,6 @@ struct ConnectionDataBundle
 {
     BluezConnection * mpConn;
     GVariant * mpVal;
-};
-
-CHIP_ERROR InitBluezBleLayer(bool aIsCentral, char * apBleAddr, BLEAdvConfig & aBleAdvConfig, void *& apEndpoint);
-bool BluezRunOnBluezThread(int (*aCallback)(void *), void * apClosure);
-bool SendBluezIndication(BLE_CONNECTION_OBJECT apConn, chip::System::PacketBufferHandle apBuf);
-bool CloseBluezConnection(BLE_CONNECTION_OBJECT apConn);
-CHIP_ERROR StartBluezAdv(BluezEndpoint * apEndpoint);
-CHIP_ERROR StopBluezAdv(BluezEndpoint * apEndpoint);
-CHIP_ERROR BluezGattsAppRegister(BluezEndpoint * apEndpoint);
-CHIP_ERROR BluezAdvertisementSetup(BluezEndpoint * apEndpoint);
-
-/// Write to the CHIP RX characteristic on the remote peripheral device
-bool BluezSendWriteRequest(BLE_CONNECTION_OBJECT apConn, chip::System::PacketBufferHandle apBuf);
-/// Subscribe to the CHIP TX characteristic on the remote peripheral device
-bool BluezSubscribeCharacteristic(BLE_CONNECTION_OBJECT apConn);
-/// Unsubscribe from the CHIP TX characteristic on the remote peripheral device
-bool BluezUnsubscribeCharacteristic(BLE_CONNECTION_OBJECT apConn);
-
-CHIP_ERROR
-StartDiscovery(BluezEndpoint * apEndpoint, BluezDiscoveryRequest aRequest = {});
-CHIP_ERROR StopDiscovery(BluezEndpoint * apEndpoint);
-
-CHIP_ERROR ConnectDevice(BluezDevice1 * apDevice);
-
-/// Iterates over available BlueZ adapters
-///
-/// Usage example:
-///
-///  AdapterIterator iterator;
-///  while (iterator.Next()) {
-///      std::cout << iterator.GetAddress() << std::endl;
-///  }
-///
-/// Data is provided through the bluez dbus interface. You can view
-/// this data in the commandline using commands such as:
-///
-///    busctl introspect org.bluez /org/bluez/hci0
-class AdapterIterator
-{
-public:
-    ~AdapterIterator();
-
-    /// Moves to the next DBUS interface.
-    ///
-    /// MUST be called before any of the 'current value' methods are
-    /// used (iterator gets initialized on the first call of Next).
-    bool Next();
-
-    // Information about the current value. Safe to call only after
-    // "Next" has returned true.
-    uint32_t GetIndex() const { return mCurrent.index; }
-    const char * GetAddress() const { return mCurrent.address.c_str(); }
-    const char * GetAlias() const { return mCurrent.alias.c_str(); }
-    const char * GetName() const { return mCurrent.name.c_str(); }
-    bool IsPowered() const { return mCurrent.powered; }
-
-private:
-    /// Sets up the DBUS manager and loads the list
-    void Initialize();
-
-    /// Loads the next value in the list.
-    ///
-    /// Returns true if a value could be loaded, false if no more items to
-    /// iterate through.
-    bool Advance();
-
-    static constexpr size_t kMaxAddressLength = 19; // xx:xx:xx:xx:xx:xx
-    static constexpr size_t kMaxNameLength    = 64;
-
-    GDBusObjectManager * mManager = nullptr; // DBus connection
-    GList * mObjectList           = nullptr; // listing of objects on the bus
-    GList * mCurrentListItem      = nullptr; // current item viewed in the list
-
-    // data valid only if Next() returns true
-    struct
-    {
-        uint32_t index;
-        std::string address;
-        std::string alias;
-        std::string name;
-        bool powered;
-    } mCurrent = { 0 };
 };
 
 } // namespace Internal
